refactor: move files into subfolder for refatoring
