+++ frontend/src/component/changeRequest/ChangeRequest/Changes/Change/ChangeOverwriteWarning/OverwriteWarning.test.tsx
@@ -1,6 +1,6 @@
 import { render } from 'utils/testRenderer';
 import { screen } from '@testing-library/react';
-import { OverwriteWarning } from './ChangeOverwriteWarning';
+import { OverwriteWarning } from './OverwriteWarning';
 import { ChangeRequestState } from 'component/changeRequest/changeRequest.types';
 
 test.each([
